Add regex filters to skip specific container log lines.
Will be used in nelm with annotation `werf.io/log-regex-skip`.

The filtering logic for the dynamic readiness tracker differs from the
multitracker, which is based on the original implementation. In the
multitracker, container-specific regexes mutually exclude global
regexes, but in the dynamic readiness, they are applied one on top of
the other. I preserved a similar logic for the skip regex. Allow users
to use both skip and regular filters to exclude parts of the logs and
then filter the results.

Closes #369

Signed-off-by: Vladimir Kuznichenkov <kuzaxak.tech@gmail.com>